Rename breadcrumb component helper
- adds `_helper` to bring it in line with convention and allow component auditing to detect it
