+++ lib/govuk_publishing_components.rb
@@ -10,7 +10,7 @@ require "govuk_publishing_components/presenters/component_wrapper_helper"
 require "govuk_publishing_components/presenters/accordion_helper"
 require "govuk_publishing_components/presenters/attachment_helper"
 require "govuk_publishing_components/presenters/big_number_helper"
-require "govuk_publishing_components/presenters/breadcrumbs"
+require "govuk_publishing_components/presenters/breadcrumbs_helper"
 require "govuk_publishing_components/presenters/breadcrumb_selector"
 require "govuk_publishing_components/presenters/button_helper"
 require "govuk_publishing_components/presenters/contextual_navigation"
